Rewrite config saving logic so less system calls are needed
Now a object, _config, mimic the file system. Object are loaded from it and from disk only when needed. a variable, _common_configs_path, save the root of the loaded configs

Ensure_file and Load_file are merged inside _load_file, the beaviour is choosed with a parameter.A set of loaded subtrees are saved in _loaded_path, and values are loaded ( without a "reload") only if they arent inside one of the subtree

pathlib is now used instead of os.path
An additional entry-point is created, ensure_file. It will take care of resolving paths before calling _load_file

Refactored config_manager to be more readable